feat: Merge branch 'feat/sparrow-refactoring' of https://github.com/sparrowapp-dev/sparrow-app into feat/create-default-page-for-collection [#834]
