+++ src/packages/@app/components/App.svelte
@@ -109,7 +109,7 @@
       <Route path="/*"><Navigate to="/app/" /></Route>
     </section>
     <section slot="unauthorized">
-      {#if 1}
+      {#if 0}
         <Route path="/init" component={EntryPoint} />
         <Route path="/*"><Navigate to="/init" /></Route>
       {:else}
