+++ src/packages/@library/icons/Vector.svelte
@@ -1,0 +1,10 @@
+<script>
+    export let height = "2px";
+    export let width = "24px";
+    export let color = "grey";
+  </script>
+
+<svg {width} {height} viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+<path d="M1.0002 2.99979H11.0002V0.999786C11.0002 0.596662 11.2439 0.231037 11.6189 0.0747866C11.9939 -0.0814634 12.4221 0.00603655 12.7096 0.290412L15.7096 3.29041C15.8971 3.47791 16.0033 3.73104 16.0033 3.99666C16.0033 4.26229 15.8971 4.51541 15.7096 4.70291L12.7096 7.70291C12.4221 7.99041 11.9939 8.07479 11.6189 7.91854C11.2439 7.76229 11.0002 7.39979 11.0002 6.99354V4.99979H1.0002C0.44707 4.99979 0.000195313 4.55291 0.000195313 3.99979C0.000195313 3.44666 0.44707 2.99979 1.0002 2.99979ZM15.0002 10.9998C15.5533 10.9998 16.0002 11.4467 16.0002 11.9998C16.0002 12.5529 15.5533 12.9998 15.0002 12.9998H5.0002V14.9998C5.0002 15.4029 4.75645 15.7685 4.38145 15.9248C4.00645 16.081 3.57832 15.9935 3.29082 15.7092L0.29082 12.7092C0.10332 12.5217 -0.00292969 12.2685 -0.00292969 12.0029C-0.00292969 11.7373 0.10332 11.4842 0.29082 11.2967L3.29082 8.29666C3.57832 8.00916 4.00645 7.92479 4.38145 8.08104C4.75645 8.23729 5.0002 8.59979 5.0002 9.00604V11.006H15.0002V10.9998Z" 
+fill={color}/>
+</svg>
