+++ src/packages/@library/icons/index.ts
@@ -14,3 +14,5 @@ export { default as FolderIcon } from "./Folder.svelte";
 export { default as SyncIcon } from "./Sync.svelte";
 export { default as GithubIcon } from "./Github.svelte";
 export { default as DoubleArrowIcon } from "./DoubleArrow.svelte";
+export { default as LibraryIcon} from "./Library.svelte";
+export { default as VectorIcon} from "./Vector.svelte";
