+++ src/packages/@workspaces/features/index.ts
@@ -6,4 +6,4 @@ export * from "./import-curl";
 export * from "./collection-explorer";
 export * from "./folder-explorer";
 export * from "./tab-bar";
-export * from "./workspace-default-page"; 
+export * from "./workspace-default";
