--- src/packages/@workspaces/features/workspace-default-page/layout/index.ts
@@ -1,1 +1,0 @@
-export { default as WorkspaceDefaultPage } from "./WorkspaceDefaultPage.svelte";
