+++ src/packages/@workspaces/features/workspace-default/layout/index.ts
@@ -1,0 +1,1 @@
+export { default as WorkspaceDefaultPage } from "./WorkspaceDefault.svelte";
